Вторая версия: числа стали вещественными

+++ main.c
@@ -6,11 +6,11 @@ int main() {
     //настройка кодировки
     system("chcp 1251");
     system("cls");
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
     printf("Введите координаты первой точки (x1, y1): ");
-    scanf("%d %d", &x1, &y1);
+    scanf("%lf %lf", &x1, &y1);
     printf("Введите координаты второй точки (x2, y2): ");
-    scanf("%d %d", &x2, &y2);
+    scanf("%lf %lf", &x2, &y2);
 
     double distance = sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
     printf("Расстояние: %.2f\n", distance);
